Move Filter option into ArticleContainer commponent
